requirements/common.txt
alembic==0.8.6
apipkg==1.4
bcrypt==3.1.0
blinker==1.4
boto==2.41.0
boto3==1.3.1
botocore==1.4.36
cffi==1.7.0
click==6.6
coverage==4.1
credstash==1.11.0
cryptography==1.4
docopt==0.6.2
docutils==0.12
ecdsa==0.13
eradicate==0.1.6
execnet==1.4.1
flake8==2.6.2
Flask==0.11.1
git+https://github.com/miracle2k/flask-assets.git@master#egg=flask-assets
Flask-Humanize==0.3.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.1
flask-oidc==1.0.4
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
future==0.15.2
httplib2==0.9.2
humanize==0.5.1
idna==2.1
itsdangerous==0.24
jinja-to-js==3.1.0
Jinja2==2.8
jmespath==0.9.0
libsass==0.11.1
Mako==1.0.4
MarkupSafe==0.23
mccabe==0.5.0
mock==2.0.0
monotonic==0.3
git+https://github.com/alphagov/notifications-python-client.git@1.0.0#egg=notifications-python-client
oauth2client==2.2.0
passlib==1.6.5
pbr==1.10.0
psycopg2==2.6.2
py==1.4.31
pyasn1==0.1.9
pycodestyle==2.0.0
pycparser==2.14
pycrypto==2.6.1
pyflakes==1.2.3
PyJWT==1.4.0
pytest==2.9.2
pytest-cache==1.0
pytest-cov==2.3.0
pytest-eradicate==0.0.3
pytest-flake8==0.5
pytest-flask==0.10.0
pytest-spec==1.0.1
pytest-splinter==1.7.3
python-dateutil==2.5.3
python-editor==1.0.1
python-jose==1.0.0
requests==2.10.0
selenium==2.53.6
six==1.10.0
splinter==0.7.3
SQLAlchemy==1.0.14
webassets==0.11.1
Werkzeug==0.11.10
WTForms==2.1
